3.0.0
The crazy update! She asked me so much time, but first of all, I would like to apologize for the last update, which had a problem EVERYWHERE, whether it was the code that was not finished, or even the comedy during which I was wrong update name, it was so buggy that mkalias did not launch.